requirements.txt
AccessControl==6.0
Acquisition==5.0
asgiref==3.6.0
attrs==22.2.0
AuthEncoding==5.0
beautifulsoup4==4.12.0
borg.localrole==3.1.10
BTrees==5.0
certifi==2022.12.7
cffi==1.15.1
Chameleon==4.0.0
channels==4.0.0
charset-normalizer==3.1.0
collective.monkeypatcher==1.2.1
cryptography==40.0.1
cssselect==1.2.0
DateTime==5.1
decorator==5.1.1
defusedxml==0.7.1
diazo==1.5.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-environ==0.10.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
DocumentTemplate==4.4
docutils==0.19
drf-spectacular==0.26.1
environ==1.0
ExtensionClass==5.0
feedparser==6.0.10
five.customerize==2.1.0
five.intid==1.2.7
five.localsitemanager==4.0
future==0.18.3
icalendar==5.0.4
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
lxml==4.9.2
Markdown==3.4.3
Missing==4.2
MultiMapping==5.0
multipart==0.2.4
oauthlib==3.2.2
Paste==3.5.2
PasteDeploy==3.0.1
Persistence==4.0.post1
persistent==5.0
piexif==1.1.3
Pillow==9.4.0
plone.alterego==1.1.5
plone.api==2.0.1
plone.app.caching==3.0.4
plone.app.content==4.0.1
plone.app.contentlisting==3.0.1
plone.app.contentmenu==3.0.0
plone.app.contentrules==5.0.0
plone.app.contenttypes==3.0.1
plone.app.customerize==1.3.12
plone.app.dexterity==3.0.4
plone.app.discussion==4.0.0
plone.app.event==4.0.0
plone.app.i18n==4.0.0
plone.app.intid==1.1.4
plone.app.iterate==5.0.1
plone.app.layout==4.0.2
plone.app.linkintegrity==4.0.0
plone.app.locales==6.0.12
plone.app.lockingbehavior==1.0.7
plone.app.multilingual==6.0.2
plone.app.portlets==5.0.2
plone.app.querystring==2.0.2
plone.app.redirector==3.0.0
plone.app.registry==2.0.1
plone.app.relationfield==3.0.1
plone.app.textfield==2.0.0
plone.app.theming==5.0.2
plone.app.upgrade==3.0.3
plone.app.users==3.0.1
plone.app.uuid==2.2.2
plone.app.versioningbehavior==2.0.1
plone.app.viewletmanager==4.0.1
plone.app.vocabularies==5.0.1
plone.app.widgets==4.0.1
plone.app.workflow==5.0.2
plone.app.z3cform==4.0.2
plone.autoform==2.0.1
plone.autoinclude==1.0.1
plone.base==1.1.0
plone.batching==2.0.4
plone.behavior==1.4.0
plone.browserlayer==3.0.1
plone.cachepurging==3.0.1
plone.caching==1.2.2
plone.contentrules==2.1.3
plone.dexterity==3.0.2
plone.event==1.4.2
plone.folder==3.1.0
plone.formwidget.namedfile==3.0.2
plone.formwidget.recurrence==3.0.3
plone.i18n==5.0.1
plone.indexer==1.0.7
plone.intelligenttext==3.1.0
plone.keyring==3.2.1
plone.locking==2.3.0
plone.memoize==3.0.0
plone.namedfile==6.0.1
plone.outputfilters==5.0.1
plone.portlet.collection==4.0.1
plone.portlet.static==4.0.1
plone.portlets==2.3.3
plone.protect==4.1.8
plone.registry==1.2.1
plone.resource==2.1.4
plone.resourceeditor==3.0.4
plone.rest==3.0.0
plone.restapi==8.35.3
plone.rfc822==2.0.2
plone.scale==4.0.1
plone.schema==1.4.0
plone.schemaeditor==4.0.3
plone.session==4.0.3
plone.staticresources==2.1.0
plone.stringinterp==2.0.0
plone.subrequest==2.0.3
plone.supermodel==2.0.3
plone.theme==3.0.7
plone.transformchain==2.0.2
plone.uuid==1.0.6
plone.volto==4.0.8
plone.z3cform==2.0.1
plonetheme.barceloneta==3.1.0
Products.BTreeFolder2==5.0
Products.CMFCore==2.7.0
Products.CMFDiffTool==4.0.0
Products.CMFDynamicViewFTI==7.0.0
Products.CMFEditions==4.0.0
Products.CMFPlacefulWorkflow==3.0.3
Products.CMFPlone==6.0.3
Products.CMFUid==4.0
Products.DateRecurringIndex==3.0.1
Products.DCWorkflow==3.0
Products.ExtendedPathIndex==4.0.1
Products.ExternalMethod==5.0
Products.GenericSetup==3.0.1
Products.isurlinportal==2.0.0
Products.MailHost==5.0
Products.MimetypesRegistry==2.1.9
Products.PlonePAS==7.0.0
Products.PluggableAuthService==2.8.1
Products.PluginRegistry==2.0
Products.PortalTransforms==4.0.1
Products.PythonScripts==5.0
Products.Sessions==4.15
Products.SiteErrorLog==6.0
Products.StandardCacheManagers==5.0
Products.statusmessages==5.0.5
Products.ZCatalog==7.0
Products.ZopeVersionControl==4.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-gettext==4.1
python3-openid==3.2.0
pytz==2023.2
pywin32==306
PyYAML==6.0
Record==4.0
repoze.xmliter==0.6.1
requests==2.28.2
requests-oauthlib==1.3.1
RestrictedPython==7.0a1.dev0
roman==4.0
sgmllib3k==1.0.0
simplejson==3.18.4
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
transaction==3.1.0
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.15
waitress==2.1.2
WebOb==1.8.7
webresource==1.2
WebTest==3.0.0
whitenoise==6.4.0
WSGIProxy2==0.5.1
z3c.caching==3.0
z3c.form==4.3
z3c.formwidget.query==1.0.0
z3c.objpath==1.3
z3c.pt==4.0
z3c.relationfield==1.0
z3c.zcmlhook==2.0
zc.lockfile==3.0.post1
zc.relation==1.2
ZConfig==3.6.1
zdaemon==4.4
ZEO==5.4.0
zExceptions==4.3
ZODB==5.8.0
ZODB3==3.11.0
zodbpickle==3.0.1
Zope==5.8.1
zope.annotation==5.0
zope.app.locales==4.3
zope.browser==3.0
zope.browsermenu==5.0
zope.browserpage==5.0
zope.browserresource==5.0
zope.cachedescriptors==5.0
zope.component==5.1.0
zope.componentvocabulary==2.3.0
zope.configuration==4.4.1
zope.container==5.0
zope.contentprovider==4.2.1
zope.contenttype==5.0
zope.copy==4.3
zope.datetime==4.3.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.dottedname==6.0
zope.event==4.6
zope.exceptions==4.6
zope.filerepresentation==6.0
zope.globalrequest==2.0
zope.hookable==5.4
zope.i18n==5.0
zope.i18nmessageid==6.0.1
zope.interface==6.0
zope.intid==5.0
zope.keyreference==5.0.0
zope.lifecycleevent==4.4
zope.location==4.3
zope.pagetemplate==5.0
zope.processlifetime==3.0
zope.proxy==5.0.0
zope.ptresource==5.0
zope.publisher==6.1.0
zope.ramcache==2.4
zope.schema==7.0.1
zope.security==6.1
zope.sendmail==5.3
zope.sequencesort==5.0
zope.site==4.6.1
zope.size==4.4
zope.structuredtext==4.4
zope.tal==5.0.1
zope.tales==5.2
zope.testbrowser==6.0
zope.testing==5.0.1
zope.traversing==5.0
zope.viewlet==5.0
Zope2==4.0
